removed the two dialog files and fixed error message occuring on delete
