+++ client/src/components/timetable/ExpandedEventView.tsx
@@ -1,16 +1,23 @@
 import React, { useContext, useState } from 'react';
-import { Dialog } from '@mui/material';
-import { daysShort } from '../../constants/timetable';
+import { AccessTime, Close, Delete, Edit, Event, LocationOn, Notes, Save } from '@mui/icons-material';
+import { Dialog, Grid, IconButton, ListItem, ListItemIcon, TextField, Typography } from '@mui/material';
+import { TimePicker } from '@mui/x-date-pickers';
+import { daysLong, daysShort } from '../../constants/timetable';
+import { AppContext } from '../../context/AppContext';
 import { CourseContext } from '../../context/CourseContext';
 import { EventTime } from '../../interfaces/Periods';
 import { ExpandedEventViewProps } from '../../interfaces/PropTypes';
+import { ExecuteButton, StyledListItem, StyledListItemText } from '../../styles/CustomEventStyles';
+import { StyledDialogContent, StyledDialogTitle, StyledTitleContainer } from '../../styles/ExpandedViewStyles';
+import { to24Hour } from '../../utils/convertTo24Hour';
+import { createNewEvent } from '../../utils/createEvent';
 import { useEventDrag } from '../../utils/Drag';
-import { createDateWithTime } from '../../utils/eventTimes';
+import { areValidEventTimes, createDateWithTime } from '../../utils/eventTimes';
+import ColorPicker from '../controls/ColorPicker';
 import DiscardDialog from './DiscardDialog';
-import DroppedEventDialog from './DroppedEventDialog';
-import EditEventDialog from './EditEventDialog';
+import DropdownOption from './DropdownOption';
 
-const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popupOpen, handleClose, isEditing, setIsEditing }) => {
+const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popupOpen, handleClose, setIsEditing, isEditing }) => {
   const { name, location, description, color } = eventPeriod.event;
   const { day, start, end } = eventPeriod.time;
 
@@ -24,9 +31,48 @@ const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popu
   const [newLocation, setNewLocation] = useState<string>(location);
   const [newDescription, setNewDescription] = useState<string>(description);
 
+  const [colorPickerAnchorEl, setColorPickerAnchorEl] = useState<HTMLElement | null>(null);
   const [newColor, setNewColor] = useState<string>(color as string);
 
   const { createdEvents, setCreatedEvents } = useContext(CourseContext);
+  const { setErrorVisibility, setAlertMsg } = useContext(AppContext);
+
+  const handleOpenColorPicker = (event: React.MouseEvent<HTMLElement>) => {
+    setColorPickerAnchorEl(event.currentTarget);
+  };
+
+  const handleCloseColorPicker = () => {
+    setColorPickerAnchorEl(null);
+  };
+
+  const handleFormat = (newFormats: string[]) => {
+    setNewDays(newFormats);
+    setIsChanged(true);
+  };
+
+  /**
+   * @param isChanged Indicates if an edit has been made to the start of the event
+   * @param newStartTime The new edited start time
+   * @param start The original starting time of the event
+   * @returns The correct time to display on the time picker.
+   * The newStartTime is only displayed if an edit has been made, otherwise the
+   * original start time is shown.
+   */
+  const timePickerStart = (isChanged: boolean, newStartTime: Date, start: number) => {
+    return isChanged ? newStartTime : createDateWithTime(start);
+  };
+
+  /**
+   * @param isChanged Indicates if an edit has been made to the end of the event
+   * @param newEndTime The new edited start time
+   * @param end The original ending time of the event
+   * @returns The correct time to display on the time picker.
+   * The newEndTime is only displayed if an edit has been made, otherwise the
+   * original end time is shown.
+   */
+  const timePickerEnd = (isChanged: boolean, newEndTime: Date, end: number) => {
+    return isChanged ? newEndTime : createDateWithTime(end);
+  };
 
   const updateEventTime = (eventTime: EventTime, id: string) => {
     setCreatedEvents({
@@ -45,6 +91,58 @@ const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popu
   };
 
   useEventDrag(updateEventTime);
+
+  const handleUpdateEvent = (id: string) => {
+    if (!areValidEventTimes(newStartTime, newEndTime)) {
+      setAlertMsg('End time is earlier than start time');
+      setErrorVisibility(true);
+      return;
+    }
+
+    // For cloning events
+    if (newDays.length > 1) {
+      createEvents(id);
+      return;
+    }
+
+    const newEventTime = {
+      day: daysShort.indexOf(newDays.toString()) + 1,
+      start: newStartTime.getHours() + newStartTime.getMinutes() / 60,
+      end: newEndTime.getHours() + newEndTime.getMinutes() / 60,
+    };
+
+    setCreatedEvents({
+      ...createdEvents,
+      [id]: {
+        type: 'event',
+        event: {
+          id: id,
+          name: newName,
+          location: newLocation,
+          description: newDescription,
+          color: newColor,
+        },
+        time: newEventTime,
+      },
+    });
+
+    setIsEditing(false);
+  };
+
+  const createEvents = (id: string) => {
+    const updatedEventData = { ...createdEvents };
+    // Delete the original event
+    delete updatedEventData[id];
+
+    // Create an event for each day that is selected in the dropdown option
+    for (const day of newDays) {
+      const newEvent = createNewEvent(newName, newLocation, newDescription, newColor, day, newStartTime, newEndTime);
+      updatedEventData[newEvent.event.id] = newEvent;
+    }
+
+    setCreatedEvents(updatedEventData);
+    setIsEditing(false);
+  };
 
   const handleDiscardChanges = () => {
     handleClose();
@@ -71,16 +169,125 @@ const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popu
     setIsChanged(false);
   };
 
+  const handleDeleteEvent = (id: string) => {
+    const updatedEventData = { ...createdEvents };
+    delete updatedEventData[id];
+    setCreatedEvents(updatedEventData);
+  };
+
   return (
     <Dialog open={popupOpen} maxWidth="sm" onClose={handleCloseDialog}>
       {isEditing ? (
-        <EditEventDialog 
-          eventPeriod={eventPeriod} 
-          handleCloseDialog={handleCloseDialog} 
-          setIsEditing={setIsEditing} 
-          isChanged={isChanged} 
-          setIsChanged={setIsChanged} 
-        />
+        <>
+          <StyledDialogTitle>
+            <StyledTitleContainer>
+              <Grid container justifyContent="flex-end" alignItems="center">
+                <IconButton aria-label="close" onClick={handleCloseDialog}>
+                  <Close />
+                </IconButton>
+              </Grid>
+            </StyledTitleContainer>
+          </StyledDialogTitle>
+          <StyledDialogContent>
+            <ListItem>
+              <ListItemIcon>
+                <Event />
+              </ListItemIcon>
+              <TextField
+                fullWidth={true}
+                id="outlined-required"
+                required
+                variant="outlined"
+                value={newName}
+                onChange={(e) => {
+                  setIsChanged(true);
+                  setNewName(e.target.value);
+                }}
+              />
+            </ListItem>
+            <ListItem>
+              <ListItemIcon>
+                <Notes />
+              </ListItemIcon>
+              <TextField
+                fullWidth={true}
+                label="Description (optional)"
+                id="outlined-required"
+                variant="outlined"
+                value={newDescription}
+                multiline
+                onChange={(e) => {
+                  setIsChanged(true);
+                  setNewDescription(e.target.value);
+                }}
+              />
+            </ListItem>
+            <ListItem>
+              <ListItemIcon>
+                <LocationOn />
+              </ListItemIcon>
+              <TextField
+                fullWidth={true}
+                id="outlined-required"
+                required
+                variant="outlined"
+                value={newLocation}
+                onChange={(e) => {
+                  setIsChanged(true);
+                  setNewLocation(e.target.value);
+                }}
+              />
+            </ListItem>
+            <ListItem>
+              <StyledListItemText primary="Start time" />
+              <TimePicker
+                value={timePickerStart(isChanged, newStartTime, start)}
+                onChange={(e) => {
+                  setIsChanged(true);
+                  if (e) setNewStartTime(e);
+                }}
+              />
+            </ListItem>
+            <ListItem>
+              <StyledListItemText primary="End time" />
+              <TimePicker
+                value={timePickerEnd(isChanged, newEndTime, end)}
+                label={!areValidEventTimes(newStartTime, newEndTime) ? 'End time must be after start' : ''}
+                slotProps={{ textField: { color: areValidEventTimes(newStartTime, newEndTime) ? 'primary' : 'error' } }}
+                onChange={(e) => {
+                  setIsChanged(true);
+                  if (e) setNewEndTime(e);
+                }}
+              />
+            </ListItem>
+            <ListItem disablePadding={true}>
+              <DropdownOption
+                optionName="Days"
+                optionState={newDays}
+                setOptionState={handleFormat}
+                optionChoices={daysShort}
+                multiple={true}
+                noOff
+              />
+            </ListItem>
+            <ColorPicker
+              color={newColor}
+              setColor={setNewColor}
+              colorPickerAnchorEl={colorPickerAnchorEl}
+              handleOpenColorPicker={handleOpenColorPicker}
+              handleCloseColorPicker={handleCloseColorPicker}
+            />
+          </StyledDialogContent>
+          <ExecuteButton
+            variant="contained"
+            color="primary"
+            onClick={() => handleUpdateEvent(eventPeriod.event.id)}
+            disabled={newName === '' || newLocation === '' || newDays.length === 0}
+          >
+            <Save />
+            SAVE
+          </ExecuteButton>
+        </>
       ) : (
         <>
           <DiscardDialog
@@ -89,12 +296,46 @@ const ExpandedEventView: React.FC<ExpandedEventViewProps> = ({ eventPeriod, popu
             setIsEditing={setIsEditing}
             setOpenSaveDialog={setOpenSaveDialog}
           />
-          <DroppedEventDialog 
-            eventPeriod={eventPeriod} 
-            handleCloseDialog={handleCloseDialog} 
-            setIsEditing={setIsEditing} 
-            isEditing={isEditing} 
-          />
+          <StyledDialogTitle>
+            <StyledTitleContainer>
+              <>{name}</>
+              <Grid container justifyContent="flex-end" alignItems="center">
+                <IconButton aria-label="edit" onClick={() => setIsEditing(true)} disabled={isEditing}>
+                  <Edit />
+                </IconButton>
+                <IconButton aria-label="delete" onClick={() => handleDeleteEvent(eventPeriod.event.id)}>
+                  <Delete />
+                </IconButton>
+                <IconButton aria-label="close" onClick={handleCloseDialog}>
+                  <Close />
+                </IconButton>
+              </Grid>
+            </StyledTitleContainer>
+          </StyledDialogTitle>
+          <StyledDialogContent>
+            {description.length > 0 && (
+              <StyledListItem>
+                <ListItemIcon>
+                  <Notes />
+                </ListItemIcon>
+                <Typography>{description}</Typography>
+              </StyledListItem>
+            )}
+            <StyledListItem>
+              <ListItemIcon>
+                <LocationOn />
+              </ListItemIcon>
+              <Typography>{location}</Typography>
+            </StyledListItem>
+            <StyledListItem>
+              <ListItemIcon>
+                <AccessTime />
+              </ListItemIcon>
+              <Typography>
+                {daysLong[day - 1]} {to24Hour(start)} {'\u2013'} {to24Hour(end)}
+              </Typography>
+            </StyledListItem>
+          </StyledDialogContent>
         </>
       )}
     </Dialog>
